Replace all bitfield types with uint32_t for consistency

+++ kernel/clock.h
@@ -12,18 +12,18 @@
 struct pit_command_repr
 {
     // bcd: 0 - 9999, disgusting
-    uint8_t bcd:     1;
+    uint32_t bcd:     1;
 
     // operating mode
     // e.g. 011: square wave
-    uint8_t mode:    3;
+    uint32_t mode:    3;
 
     // access mode
     // lobyte only, hibyte only, lobyte + hibyte
-    uint8_t rw:      2;
+    uint32_t rw:      2;
 
     // channel 0 - 2
-    uint8_t channel: 2;
+    uint32_t channel: 2;
 }__attribute__((packed));
 
 void clock_set_interval(int hz);
